Revert "Error de archivos js corregido y Publicaciones ya tiene plantilla"
This reverts commit 9c6722aab5afd71d91ed3f52f7bac2ced6d5804d.
