+++ public/assets/css/responsive.css
@@ -107,7 +107,6 @@
 	}
 	.site-logo a img {
 	    max-width: 150px;
-		height: 10px;
 	}
 	.site-logo {
 	    text-align: center;
