+++ public/assets/css/scss/base/_mixins.scss
@@ -42,8 +42,8 @@
 
 // Hover State
 @mixin boxed-btn-hvr {
-	background-color: #0D004B;
-    color: #FFDE59;
+	background-color: #051922;
+    color: #F28123;
 	color: $white-color;
 	background-color: darken($primary-color, $darken);
 }
