+++ public/assets/css/scss/base/_variables.scss
@@ -1,7 +1,7 @@
 // Colors
 
-$primary-color: #FFDE59;
-$text-color: #0D004B;
+$primary-color: #F28123;
+$text-color: #051922;
 $text-color-light: #999999;
 $white-color: #fff;
 $black-color: #000;
