+++ public/assets/css/scss/common/_header.scss
@@ -236,7 +236,7 @@ ul.navbar-nav {
   left: 0;
   top: 0;
   z-index: 5555;
-  background-color: #0D004B;
+  background-color: #051922;
   width: 100%;
   height: 100%;
   text-align: center;
