+++ public/assets/css/scss/sections/_cart-banner.scss
@@ -18,7 +18,7 @@
       height: 110px;
       border-radius: 50%;
       border: 1px solid $white-color;
-      background-color: rgba(255, 189, 89, 0.75);
+      background-color: rgba(242, 129, 35, 0.75);
 
       .inner-price {
         position: relative;
