+++ src/components/Navbar.jsx
@@ -13,7 +13,7 @@ function Navbar(){
                         <div className="main-menu-wrap">
                         <div className="site-logo">
                             <NavLink to="/">
-                            <img src="assets/img/logo.png" alt="" style={{ width: 'auto', height: '50px' }}/>
+                            <img src="assets/img/logo.png" alt="" />
                             </NavLink>
                         </div>
 
@@ -35,16 +35,9 @@ function Navbar(){
                                 </NavLink>
                             </li>
 
-                            <li>
-                                {/* Links de sesión */}
-                                {usuario && (
-                                    <span style={{ color: '#FFDE59' }}>
-                                        {usuario.nombre}
-                                        {usuarioAuth && usuarioAuth.emailVerified ? " - Verificado" : " - No verificado"}
-                                    </span>
-                                )}
-                            </li>
-
+                            {/* Links de sesión */}
+                            {usuario && usuario.nombre}
+                            {usuario && (usuarioAuth && usuarioAuth.emailVerified ? "Verificado" : "No verificado")}
                             <li>
                                 {usuario ? (
                                 <NavLink to="/" onClick={cerrarSesion}>
@@ -65,22 +58,6 @@ function Navbar(){
                         <div className="mobile-menu"></div>
                         </div>
                     </div>
-                    </div>
-                </div>
-            </div>
-            <div className="search-area">
-                <div className="container">
-                    <div className="row">
-                        <div className="col-lg-12">
-                            <span className="close-btn"><i className="fas fa-window-close"></i></span>
-                            <div className="search-bar">
-                                <div className="search-bar-tablecell">
-                                    <h3>Search For:</h3>
-                                    <input type="text" placeholder="Keywords"/>
-                                    <button type="submit">Search <i className="fas fa-search"></i></button>
-                                </div>
-                            </div>
-                        </div>
                     </div>
                 </div>
             </div>
