+++ src/pages/Inicio.jsx
@@ -1,84 +1,42 @@
-import LoadScripts from '../components/LoadScripts';
-
 function Inicio(){
-
+    
     return(
         <div>
-        {LoadScripts()}
-
-        {/* home page slider */}
-        <div className="homepage-slider">
-            {/* single home slider */}
-            <div className="single-homepage-slider homepage-bg-1">
+        <div className="search-area">
             <div className="container">
                 <div className="row">
-                <div className="col-md-12 col-lg-7 offset-lg-1 offset-xl-0">
-                    <div className="hero-text">
-                    <div className="hero-text-tablecell">
-                        <p className="subtitle">Fresh &amp; Organic</p>
-                        <h1>Delicious Seasonal Fruits</h1>
-                        <div className="hero-btns">
-                        <a href="shop.html" className="boxed-btn">
-                            Fruit Collection
-                        </a>
-                        <a href="contact.html" className="bordered-btn">
-                            Contact Us
-                        </a>
+                    <div className="col-lg-12">
+                        <span className="close-btn"><i className="fas fa-window-close"></i></span>
+                        <div className="search-bar">
+                            <div className="search-bar-tablecell">
+                                <h3>Search For:</h3>
+                                <input type="text" placeholder="Keywords"/>
+                                <button type="submit">Search <i className="fas fa-search"></i></button>
+                            </div>
                         </div>
                     </div>
-                    </div>
                 </div>
-                </div>
-            </div>
-            </div>
-            {/* single home slider */}
-            <div className="single-homepage-slider homepage-bg-2">
-            <div className="container">
-                <div className="row">
-                <div className="col-lg-10 offset-lg-1 text-center">
-                    <div className="hero-text">
-                    <div className="hero-text-tablecell">
-                        <p className="subtitle">Fresh Everyday</p>
-                        <h1>100% Organic Collection</h1>
-                        <div className="hero-btns">
-                        <a href="shop.html" className="boxed-btn">
-                            Visit Shop
-                        </a>
-                        <a href="contact.html" className="bordered-btn">
-                            Contact Us
-                        </a>
-                        </div>
-                    </div>
-                    </div>
-                </div>
-                </div>
-            </div>
-            </div>
-            {/* single home slider */}
-            <div className="single-homepage-slider homepage-bg-3">
-            <div className="container">
-                <div className="row">
-                <div className="col-lg-10 offset-lg-1 text-right">
-                    <div className="hero-text">
-                    <div className="hero-text-tablecell">
-                        <p className="subtitle">Mega Sale Going On!</p>
-                        <h1>Get December Discount</h1>
-                        <div className="hero-btns">
-                        <a href="shop.html" className="boxed-btn">
-                            Visit Shop
-                        </a>
-                        <a href="contact.html" className="bordered-btn">
-                            Contact Us
-                        </a>
-                        </div>
-                    </div>
-                    </div>
-                </div>
-                </div>
-            </div>
             </div>
         </div>
-        {/* end home page slider */}
+
+        <div className="hero-area hero-bg">
+            <div className="container">
+                <div className="row">
+                    <div className="col-lg-9 offset-lg-2 text-center">
+                        <div className="hero-text">
+                            <div className="hero-text-tablecell">
+                                <div className="subtitle"><p>Fresh & Organic</p></div>
+                                <h1>Delicious Seasonal Fruits</h1>
+                                <div className="hero-btns">
+                                    <a href="shop.html" className="boxed-btn">Fruit Collection</a>
+                                    <a href="contact.html" className="bordered-btn">Contact Us</a>
+                                </div>
+                            </div>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </div>
 	
         <div className="list-section pt-80 pb-80">
             <div className="container">
@@ -187,7 +145,7 @@ function Inicio(){
                         <h3><span className="orange-text">Deal</span> of the month</h3>
                         <h4>Hikan Strwaberry</h4>
                         <div className="text">Quisquam minus maiores repudiandae nobis, minima saepe id, fugit ullam similique! Beatae, minima quisquam molestias facere ea. Perspiciatis unde omnis iste natus error sit voluptatem accusant</div>
-                        {/*Countdown Timer*/}
+			    {/*Countdown Timer*/}
                         <div className="time-counter"><div className="time-countdown clearfix" data-countdown="2020/2/01"><div className="counter-column"><div className="inner"><span className="count">00</span>Days</div></div> <div className="counter-column"><div className="inner"><span className="count">00</span>Hours</div></div>  <div className="counter-column"><div className="inner"><span className="count">00</span>Mins</div></div>  <div className="counter-column"><div className="inner"><span className="count">00</span>Secs</div></div></div></div>
                         <a href="cart.html" className="cart-btn mt-3"><i className="fas fa-shopping-cart"></i> Add to Cart</a>
                     </div>
